+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change log for LDAPCP
 
+## Unreleased
+
+* Fix error when creating the configuration, due to case-sensitive test in the claim types - https://github.com/Yvand/LDAPCP/issues/204
+* Fix the error when loading the global configuration page, if the group claim type set in the LDAPCP configuration does not exist in the trust - https://github.com/Yvand/LDAPCP/issues/203
+* Add the property MaxSearchResultsCount, to override the SharePoint limit of the maximum number of objects that the LDAP server returns - https://github.com/Yvand/LDAPCP/issues/209
+* Correctly initialize LDAP-specific properties with their actual value, instead of the default value of the type - https://github.com/Yvand/LDAPCP/pull/212
+* Fix an NullReferenceException in a very rare scenario where ClaimsPrincipal.Identity is null
+* Add helper methods to get/delete a directory connection in the configuration
+
 ## LDAPCP Second Edition v17.0.20240226.2 - Published in February 26, 2024
 
-* Initial release of LDAPCP Second Edition, a complete rewrite of current project
+* Ignore case when comparing claim types, to avoid errors when creating the configuration - https://github.com/Yvand/LDAPCP/pull/205
 
 ## LDAPCP v16.0.20230824.1 enhancements & bug-fixes - Published in August 24, 2023